feat: add error handling for user services

+++ internal/service/user.go
@@ -4,39 +4,48 @@ import (
 	"context"
 	"creepy/internal/models"
 	"creepy/internal/storage"
+	"fmt"
 
 	"github.com/google/uuid"
 )
 
 type UserService struct {
-    Repo storage.UserRepository
+	Repo storage.UserRepository
 }
 
 func NewUserService(repo storage.UserRepository) *UserService {
-	// TODO: Error handling
-    return &UserService{Repo: repo}
+	return &UserService{Repo: repo}
 }
 
 func (s *UserService) CreateUser(ctx context.Context, user *models.User) error {
-	// TODO: Error handling
-    return s.Repo.Save(ctx, user)
+	if user.UserName == "" {
+		return fmt.Errorf("username can't be empty")
+	}
+	return s.Repo.Save(ctx, user)
 }
 
 func (s *UserService) GetUser(ctx context.Context, id uuid.UUID) (*models.User, error) {
-    // TODO: Error handling
-    return s.Repo.GetByID(ctx, id)
+	return s.Repo.GetByID(ctx, id)
 }
 
 func (s *UserService) UpdateUser(ctx context.Context, user *models.User) error {
-	// TODO: Error handling
-	return s.Repo.Update(ctx, user)
+	_, err := s.GetUser(ctx, user.ID)
+	if err == nil {
+		return s.Repo.Update(ctx, user)
+	} else {
+		return fmt.Errorf("this user id doesn't  exist")
+	}
 }
 
 func (s *UserService) DeleteUser(ctx context.Context, id uuid.UUID) error {
-	// TODO: Error handling
-	return s.Repo.Delete(ctx, id)
+	_, err := s.GetUser(ctx, id)
+	if err == nil {
+		return s.Repo.Delete(ctx, id)
+	} else {
+		return fmt.Errorf("invalid id")
+	}
 }
 
 func (s *UserService) GetByUserName(ctx context.Context, userName string) (*models.User, error) {
-    return s.Repo.GetByUserName(ctx, userName)
+	return s.Repo.GetByUserName(ctx, userName)
 }
